Moved the disk tests to a separate file
