+++ src/test/java/CRUMTest.java
@@ -26,7 +26,6 @@ public class CRUMTest {
       assertNotNull(crum.hal);
       assertNotNull(crum.diskData);
       assertNotNull(crum.SerialNum);
-      //assertNotNull(crum.fs);
       assertNotNull(crum.numDisks);
       assertNotNull(crum.CPUdata);
       assertNotNull(crum.memoryData);
@@ -69,25 +68,6 @@ public class CRUMTest {
       c.close();
   }
 
-/*    @Test
-    public void initDiscTest() throws SQLException {
-        CRUM crum = new CRUM();
-        crum.initOSHI();
-        crum.initDB();
-        crum.initMachine();
-        c = DriverManager.getConnection("jdbc:sqlite:crum.db");
-        Calendar calendar = Calendar.getInstance();
-        crum.getDiskData(calendar);
-        stmt = c.createStatement();
-        String sql_Search = "SELECT * FROM DISC ";
-        ResultSet rs = stmt.executeQuery(sql_Search);
-        assertEquals(0, rs.getInt("DISC_ID"));
-        assertEquals(crum.fileStores.get(0).getMount(), rs.getString("DISC_NAME"));
-        assertEquals((crum.fileStores.get(0).getTotalSpace()/1000000000), rs.getLong("DISC_SIZE"));
-        c.close();
-    }
- */
-
     /*
     @Test
     public void initNetworkTest() throws SQLException {
@@ -128,54 +108,6 @@ public class CRUMTest {
         assertEquals(Macs, rs.getString("MAC_ADDRESS"));
         c.close();
     }
-
      */
-
- /*    @Test
-    void getDiskDataTest(){
-        CRUM crum = new CRUM();
-        Calendar calendar = Calendar.getInstance();
-        crum.initOSHI();
-        crum.initDB();
-        crum.getDiskData(calendar);
-        String machineID = crum.hal.getComputerSystem().getSerialNumber();
-        String sourceFile = "test.txt";
-        try {
-            c = DriverManager.getConnection("jdbc:sqlite:crum.db");
-            stmt = c.createStatement();
-            String sql_Search = "SELECT * FROM DISC ";
-            ResultSet rs = stmt.executeQuery(sql_Search);
-            int initUsed = rs.getInt("DISC_USED");
-            OutputStream os = new FileOutputStream("test2.txt");
-            InputStream is = new FileInputStream(sourceFile);
-            os.write(is.read());
-            c.close();
-            is.close();
-            os.close();
-            calendar = Calendar.getInstance();
-            crum.getDiskData(calendar);
-            c = DriverManager.getConnection("jdbc:sqlite:crum.db");
-            stmt = c.createStatement();
-            java.sql.Timestamp currentTime = new java.sql.Timestamp(calendar.getTime().getTime());
-            String timeSearch = currentTime.toString();
-            String[] timeSplitTemp = timeSearch.split(".");
-            timeSearch = timeSplitTemp[0];
-            String secondSearch = "SELECT * FROM DISC WHERE DATETIME(TIMESTAMP) >= '" + timeSearch + "'";
-            rs = stmt.executeQuery(secondSearch);
-            int finalUsed = rs.getInt("DISC_USED");
-            assertTrue((finalUsed > initUsed));
-            c.close();
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
-        } catch (IOException e) {
-            e.printStackTrace();
-        } catch (SQLException throwables) {
-            throwables.printStackTrace();
-        } catch (Exception e){
-
-        }
-    }
-
-  */
 
 }
